simplified nlog to be ilogger rather than fulllogger
